// core/models/Inventario/Categoria.ts
export interface Categoria {
    id: number;
    nombre: string;
    equipo: string;
    suspendido: boolean;
    otrosF1: string;
    otrosN1: number;
    otrosN2: number;
    otrosC1: null;
    otrosC2: null;
    otrosC3: null;
    otrosC4: null;
    otrosT1: null;
    fechaRegistro: string;
    usuarioRegistroNombre: string;
    fechaModificacion: string;
    usuarioModificacionNombre: string;
  }
  